Pin integer widths and use COO matrix layout for block-sparse coordinates
Cross-language readability changes to the on-disk Serialized* layouts:
- version :: UInt32 (matches Base.VersionNumber's field width).
- block_indices :: Matrix{Int64} shaped (ndims, num_blocks), the COO
  convention used by Apache Arrow, PyData Sparse, and PyTorch sparse.
  ndims is implicit in size(block_indices, 1) and is preserved by HDF5
  even when num_blocks == 0, so the standalone ndims field is no longer
  needed and has been removed.
- block_offsets :: Vector{Int64}.
- SerializedEmptyStorage now just carries version; the eltype is already
  encoded in the parametric type record on disk, matching how the other
  Serialized* storage types handle their eltype.

The block-offset helpers take Val(N) so the constructed BlockOffsets{N}
stays type-stable.

+++ NDTensors/ext/NDTensorsJLD2Ext/dense.jl
@@ -4,7 +4,7 @@ using NDTensors: NDTensors, Dense, SerializedDense
 JLD2.writeas(::Type{<:Dense{T}}) where {T} = SerializedDense{T}
 
 function JLD2.wconvert(::Type{SerializedDense{T}}, d::Dense{T}) where {T}
-    version = 1
+    version = UInt32(1)
     return SerializedDense{T}(version, convert(Vector{T}, NDTensors.data(d)))
 end
